py2/py3: Add futurize stage 2 changes.
These were autogenerated with "futurize -2wn .", manual fixes follow.

+++ gooey/python_bindings/docopt_to_json.py
@@ -41,7 +41,7 @@ class MyOption(Option):
 
   @classmethod
   def parse(class_, option_description):
-      short, long, argcount, value = None, None, 0, False
+      short, int, argcount, value = None, None, 0, False
       options, _, description = option_description.strip().partition('  ')
       options = options.replace(',', ' ').replace('=', ' ')
       for s in options.split():
@@ -54,7 +54,7 @@ class MyOption(Option):
       if argcount:
           matched = re.findall('\[default: (.*)\]', description, flags=re.I)
           value = matched[0] if matched else None
-      return class_(short, long, argcount, value, description=description.strip())
+      return class_(short, int, argcount, value, description=description.strip())
 
   def __repr__(self):
       return 'Option(%r, %r, %r, %r, %r)' % (self.short, self.long,
